Split ComfyUISettingsScreen into section composables (#933)
