+++ androidApp/src/main/kotlin/com/riox432/civitdeck/ui/comfyui/ComfyUIConnectionSection.kt
@@ -1,0 +1,137 @@
+package com.riox432.civitdeck.ui.comfyui
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.Card
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.RadioButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.riox432.civitdeck.R
+import com.riox432.civitdeck.domain.model.ComfyUIConnection
+import com.riox432.civitdeck.domain.model.DiscoveredServer
+import com.riox432.civitdeck.feature.comfyui.presentation.ComfyUISettingsUiState
+import com.riox432.civitdeck.ui.theme.CornerRadius
+import com.riox432.civitdeck.ui.theme.Spacing
+
+@Composable
+internal fun ScanLanSection(
+    state: ComfyUISettingsUiState,
+    onScan: () -> Unit,
+    onSelect: (DiscoveredServer) -> Unit,
+) {
+    Card(modifier = Modifier.fillMaxWidth()) {
+        Column(modifier = Modifier.padding(Spacing.md)) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                Text(stringResource(R.string.comfyui_lan_discovery), style = MaterialTheme.typography.titleSmall)
+                if (state.isScanning) {
+                    CircularProgressIndicator(modifier = Modifier.size(24.dp))
+                } else {
+                    OutlinedButton(onClick = onScan) { Text(stringResource(R.string.comfyui_scan_lan)) }
+                }
+            }
+            if (state.discoveredServers.isNotEmpty()) {
+                Spacer(modifier = Modifier.height(Spacing.sm))
+                state.discoveredServers.forEach { server ->
+                    DiscoveredServerRow(server, onSelect)
+                }
+            } else if (!state.isScanning) {
+                val scanError = state.scanError
+                if (scanError != null) {
+                    Text(
+                        stringResource(R.string.comfyui_scan_failed, scanError),
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.error,
+                    )
+                } else {
+                    Text(
+                        stringResource(R.string.comfyui_lan_scan_hint),
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant,
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+private fun DiscoveredServerRow(server: DiscoveredServer, onSelect: (DiscoveredServer) -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable { onSelect(server) }
+            .padding(vertical = Spacing.xs),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Column(modifier = Modifier.weight(1f)) {
+            Text(server.displayName, style = MaterialTheme.typography.bodyMedium)
+            Text("${server.ip}:${server.port}", style = MaterialTheme.typography.bodySmall)
+        }
+        TextButton(onClick = { onSelect(server) }) { Text(stringResource(R.string.action_add)) }
+    }
+}
+
+@Composable
+internal fun ConnectionCard(
+    connection: ComfyUIConnection,
+    isActive: Boolean,
+    onActivate: () -> Unit,
+    onEdit: () -> Unit,
+    onDelete: () -> Unit,
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable(onClick = onActivate, onClickLabel = "Select connection"),
+        shape = RoundedCornerShape(CornerRadius.card),
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth().padding(Spacing.md),
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            RadioButton(selected = isActive, onClick = onActivate)
+            Column(modifier = Modifier.weight(1f)) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Text(connection.name, style = MaterialTheme.typography.titleSmall)
+                    if (connection.isSecure) {
+                        Text(
+                            " HTTPS",
+                            style = MaterialTheme.typography.labelSmall,
+                            color = MaterialTheme.colorScheme.primary,
+                        )
+                    }
+                }
+                Text(connection.baseUrl, style = MaterialTheme.typography.bodySmall)
+            }
+            IconButton(onClick = onEdit) { Icon(Icons.Default.Edit, "Edit") }
+            IconButton(onClick = onDelete) {
+                Icon(Icons.Default.Delete, "Delete", tint = MaterialTheme.colorScheme.error)
+            }
+        }
+    }
+}
